revises team and bar models, references team into bar, cleans seeds files
