+++ db/bar-seeds.js
@@ -2,7 +2,6 @@ var mongoose = require('mongoose'),
         conn = mongoose.connect('mongodb://localhost/HCA-app'),
          Bar = require("../models/bar");
 
-
 Bar.remove({}, function(err){
   if (err) console.log("ERROR: ", err);
 });
@@ -14,15 +13,18 @@ var bars = [
     city: "SF",
     state: "CA",
     lat: 37.790882,
-    long: -122.401552
+    long: -122.401552,
+    teams: ['56e8519b7d4d581f480f8e4e']
   },
-    {
+  {
     name: "fear garden",
     address: "666 death ave",
-    city: "Oakland",
+    city: "San Francisco",
     state: "CA",
     lat: 37.790230,
-    long: -122.408439
+    long: -122.408439,
+    teams: ['56e8519b7d4d581f480f8e4f', 
+            '56e8519b7d4d581f480f8e50']
   }
 ];
