+++ views/users/show.hbs
@@ -1,6 +1,6 @@
 <h1>This is a User.</h1>
 <p>{{user}}</p>
-<p>{{user.email}}</p>
+<p>{{user.fb.email}}</p>
 <p>{{user.about}}</p>
 <p>{{user.picture}}</p>
 <p>{{user.favorites}}</p>
