+++ views/users/user-edit-form.hbs
@@ -7,7 +7,7 @@
       <input class="form-control" id="user-email" name="email" type="text" value="{{user.email}}" placeholder="email" />
     </div> --}}
     <div class="form-group col-md-4 form-centered">
-      <h3 id="user-email">Email: {{user.email}}</h3>
+      <h3 id="user-email">Email: {{user.fb.email}}</h3>
     </div>
     <div class="form-group col-md-4 form-centered">
       <textarea class="form-control" id="user-about" name="about" type="text" value="{{user.about}}" placeholder="write something about yourself">{{user.about}}</textarea>
